Web Scraper Pro is a feature rich tool that offers BeautifulSoup and Selenium as the scraping methods.
Web Scraper Pro is a feature rich tool that offers BeautifulSoup and Selenium as the scraping methods.
The GUI includes a drop-down menu, and text box that allows users to manually input a Selenium expression.
Unique Functionality supports scraping using predefined tags with BeautifulSoup and Selenium, along with the ability to save the scraped data in various formats.